Extract AbstractAttributeCollector base class, improve documentation
- Extract shared AbstractAttributeCollector to eliminate duplicate
  XMLValidator stub implementations in SimpleOutputElement and
  NonNsStreamWriter (~100 lines of boilerplate removed)
- Improve mVldException Javadoc explaining the "poisoning" semantics
  and why it was introduced (issues #179, #190)
- Add Javadoc on setValidator() explaining the ancestor recursion
- Add comments on mCheckAttrs gate in copyStartElement() methods
  clarifying it is guaranteed true when validation is active
- Add comment in _finishDocument() explaining why auto-closing is
  skipped after a validation error
